test(gpu): promote FakeGpuDetector to shared tests/_fakes helper
